requirements.txt
streamlit
folium
streamlit-folium
pandas
python-dotenv
supabase
